# api/openapi.yaml
openapi: 3.0.3
info:
  title: Prime Terrain API
  description: |
    API for exploring prime neighborhoods with cryptographic signing,
    usage tracking, and multiple export formats.

    ## Mathematical Model
    - **Gap**: g(n) = p(n+1) − p(n)
    - **Second difference**: Δ² = g(n) − g(n−1)
    - **Second ratio**: r = Δ² / (p(n+1) − p(n−1)), always in [-1, 1]

    ## Authentication
    API keys are passed via headers:
    - `Authorization: Bearer frc_...` (preferred)
    - `X-API-Key: frc_...`

    Unauthenticated requests use free tier limits.

    ## Rate Limits
    | Tier | CPU/Month | Req/Min | Expensive/Min | Async Jobs |
    |------|-----------|---------|---------------|------------|
    | free | 5 min | 100 | 20 | 5 |
    | pro | unlimited | 500 | 100 | 50 |
    | enterprise | custom | 2000 | 500 | 500 |
  version: 1.0.0
  contact:
    name: Fractal Core
    url: https://fractal-core.com
  license:
    name: MIT
servers:
  - url: https://api.fractal-core.com/api/v1
    description: Production
  - url: https://prime-terrain-api.tem-527.workers.dev/api/v1
    description: Workers Dev
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key in format frc_<32 alphanumeric chars>
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative API key header
  schemas:
...
    NType:
      type: string
      enum: [index, value]
      default: index
      description: How to interpret anchor value
    Mode:
      type: string
      enum: [count, span]
      default: count
      description: Size determination method
    Engine:
      type: string
      enum: [auto, sieve, mr64, bpsw, mr-prob, precomputed]
      default: auto
    ValidationMode:
      type: string
      enum: [none, dual, triple]
      default: none
    ProofLevel:
      type: string
      enum: [none, receipt, signed]
      default: none
    OutputFormat:
      type: string
      enum: [json, bin]
      default: json
    Compression:
      type: string
      enum: [none, gzip, zstd]
      default: none
    IncludeField:
      type: string
      enum: [primes, gaps, d2, ratio, indices]
...
    JobStatus:
      type: string
      enum: [pending, running, completed, failed]
...
    Rational:
      type: object
      required: [num, den]
      properties:
        num:
          type: string
          description: Numerator as bigint string
        den:
          type: string
          description: Denominator as bigint string
      example:
        num: "-2"
        den: "12"
    CanonicalRequest:
      type: object
      required: [n, n_type, mode, include, engine, validate, proof, format, compress]
      properties:
        n:
          type: string
          description: Anchor value as bigint string
        n_type:
          $ref: '#/components/schemas/NType'
        mode:
          $ref: '#/components/schemas/Mode'
        k:
          type: integer
          minimum: 1
          maximum: 10000
        w:
          type: string
          description: Half-width as bigint string (span mode)
        include:
          type: array
          items:
            $ref: '#/components/schemas/IncludeField'
        engine:
          $ref: '#/components/schemas/Engine'
        validate:
          $ref: '#/components/schemas/ValidationMode'
        proof:
          $ref: '#/components/schemas/ProofLevel'
        format:
          $ref: '#/components/schemas/OutputFormat'
        compress:
          $ref: '#/components/schemas/Compression'
    NeighborhoodResult:
      type: object
      required: [n, n_type, mode]
      properties:
        n:
          type: string
        n_type:
          $ref: '#/components/schemas/NType'
        mode:
          $ref: '#/components/schemas/Mode'
        center_prime:
          type: string
          description: Center prime as bigint string
        primes:
          type: array
          items:
            type: string
          description: Prime values as bigint strings
        gaps:
          type: array
          items:
            type: string
          description: Gap values as unsigned bigint strings
        d2:
          type: array
          items:
            type: string
          description: Second differences as signed bigint strings
        ratio:
          type: array
          items:
            $ref: '#/components/schemas/Rational'
        indices:
          type: array
          items:
            type: integer
    Receipt:
      type: object
      required: [request_hash, result_hash, engines, validation, deterministic, generated_at]
      properties:
        request_hash:
          type: string
          description: SHA-256 hex of canonical request
        result_hash:
          type: string
          description: SHA-256 hex of canonical result
        engines:
          type: array
          items:
            type: string
        validation:
          type: object
          required: [mode, agreement]
          properties:
            mode:
              type: string
            agreement:
              type: boolean
        deterministic:
          type: boolean
        generated_at:
          type: string
          format: date-time
    SignatureBlock:
      type: object
      required: [alg, key_id, signed_hash_alg, sig_b64]
      properties:
        alg:
          type: string
          enum: [ed25519]
        key_id:
          type: string
        signed_hash_alg:
          type: string
          enum: [sha256]
        sig_b64:
          type: string
          description: Base64-encoded Ed25519 signature
    SignedResponse:
      type: object
      required: [request, result]
      properties:
        request:
          $ref: '#/components/schemas/CanonicalRequest'
        result:
          $ref: '#/components/schemas/NeighborhoodResult'
        receipt:
          $ref: '#/components/schemas/Receipt'
        signature:
          $ref: '#/components/schemas/SignatureBlock'

    # User Types
...
    Job:
      type: object
      properties:
        job_id:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        result_url:
          type: string
        error:
          type: string

    # Error
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  parameters:
    N:
      name: n
      in: query
      required: true
      schema:
        type: string
      description: Anchor value as bigint string
    NType:
      name: n_type
      in: query
      schema:
        $ref: '#/components/schemas/NType'
    Mode:
      name: mode
      in: query
      schema:
        $ref: '#/components/schemas/Mode'
    K:
      name: k
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 10
      description: Number of primes each side (count mode)
    W:
      name: w
      in: query
      schema:
        type: string
      description: Half-width as bigint string (span mode)
    Include:
      name: include
      in: query
      schema:
        type: string
      description: Comma-separated fields (primes,gaps,d2,ratio,indices)
    Engine:
      name: engine
      in: query
      schema:
        $ref: '#/components/schemas/Engine'
    Validate:
      name: validate
      in: query
      schema:
        $ref: '#/components/schemas/ValidationMode'
    Proof:
      name: proof
      in: query
      schema:
        $ref: '#/components/schemas/ProofLevel'
    Format:
      name: format
      in: query
      schema:
        $ref: '#/components/schemas/OutputFormat'
    Compress:
      name: compress
      in: query
      schema:
        $ref: '#/components/schemas/Compression'
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until retry
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate limit exceeded
            message: Too many requests, please try again later.
            retry_after: 45
paths:
...
  /neighborhood:
    get:
      tags: [Neighborhood]
      summary: Full prime neighborhood
      description: |
        Computes prime neighborhood around anchor with all derived sequences.
        Supports cryptographic signing and multiple output formats.
      operationId: getNeighborhood
      security:
        - {}
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/N'
        - $ref: '#/components/parameters/NType'
        - $ref: '#/components/parameters/Mode'
        - $ref: '#/components/parameters/K'
        - $ref: '#/components/parameters/W'
        - $ref: '#/components/parameters/Include'
        - $ref: '#/components/parameters/Engine'
        - $ref: '#/components/parameters/Validate'
        - $ref: '#/components/parameters/Proof'
        - $ref: '#/components/parameters/Format'
        - $ref: '#/components/parameters/Compress'
      responses:
        '200':
          description: Neighborhood data
          headers:
            X-CPU-Time-Ms:
              schema:
                type: integer
            X-CPU-Time-Remaining-Ms:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List all jobs
      operationId: listJobs
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'

    post:
      tags: [Jobs]
      summary: Create async job
      description: Creates job for large computations that exceed sync limits
      operationId: createJob
      security:
        - {}
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [n]
              properties:
                n:
                  type: string
                n_type:
                  $ref: '#/components/schemas/NType'
                mode:
                  $ref: '#/components/schemas/Mode'
                k:
                  type: integer
                w:
                  type: string
                include:
                  type: array
                  items:
                    $ref: '#/components/schemas/IncludeField'
                engine:
                  $ref: '#/components/schemas/Engine'
                validate:
                  $ref: '#/components/schemas/ValidationMode'
                proof:
                  $ref: '#/components/schemas/ProofLevel'
      responses:
        '202':
          description: Job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    $ref: '#/components/schemas/JobStatus'
                  poll_url:
                    type: string
        '429':
          description: Async job limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  active_jobs:
                    type: integer
                  limit:
                    type: integer
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Get job status
      operationId: getJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found

    delete:
      tags: [Jobs]
      summary: Delete job
      operationId: deleteJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        '404':
          description: Job not found
  /jobs/{job_id}/result:
    get:
      tags: [Jobs]
      summary: Get job result
      operationId: getJobResult
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedResponse'
        '400':
          description: Job not completed
        '404':
          description: Job not found

  # OEIS
...
  /oeis/neighborhood:
    get:
      tags: [OEIS]
      summary: Full neighborhood in OEIS format
      operationId: getOeisNeighborhood
      parameters:
        - name: n
          in: query
          required: true
          schema:
            type: integer
        - name: k
          in: query
          schema:
            type: integer
            default: 10
        - name: format
          in: query
          schema:
            type: string
            enum: [json, bfile, internal, list]
            default: json
      responses:
        '200':
          description: Neighborhood in OEIS format
        '429':
          $ref: '#/components/responses/RateLimitExceeded'

  # Formats
...